add busy in hashtag channel voice
