+++ apps/mobile/src/app/components/Suggestions/SuggestItem/index.tsx
@@ -1,7 +1,11 @@
 import { useTheme } from '@mezon/mobile-ui';
-import { Image, Text, View } from 'react-native';
-import { style } from './SuggestItem.styles';
+import { selectVoiceChannelMembersByChannelId } from '@mezon/store-mobile';
 import { getSrcEmoji } from '@mezon/utils';
+import { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Image, Text, View } from 'react-native';
+import { useSelector } from 'react-redux';
+import { style } from './SuggestItem.styles';
 
 type SuggestItemProps = {
 	avatarUrl?: string;
@@ -11,12 +15,23 @@ type SuggestItemProps = {
 	isDisplayDefaultAvatar?: boolean;
 	isRoleUser?: boolean;
 	emojiId?: string;
+	channelId?: string
 };
 
-const SuggestItem = ({ avatarUrl, symbol, name, subText, isDisplayDefaultAvatar, isRoleUser, emojiId }: SuggestItemProps) => {
+const SuggestItem = ({ channelId, avatarUrl, symbol, name, subText, isDisplayDefaultAvatar, isRoleUser, emojiId }: SuggestItemProps) => {
 	const { themeValue } = useTheme();
 	const styles = style(themeValue);
 	const emojiSrc = emojiId ? getSrcEmoji(emojiId) : '';
+	const voiceChannelMember = useSelector(selectVoiceChannelMembersByChannelId(channelId));
+	const { t } = useTranslation(['clan']);
+
+	const checkVoiceStatus = useMemo(() => {
+		if (channelId !== undefined && voiceChannelMember) {
+			return voiceChannelMember.length >= 2;
+		}
+		return false;
+	}, [voiceChannelMember, channelId]);
+
 	return (
 		<View style={styles.wrapperItem}>
 			<View style={styles.containerItem}>
@@ -41,7 +56,15 @@ const SuggestItem = ({ avatarUrl, symbol, name, subText, isDisplayDefaultAvatar,
 				{isRoleUser || name.startsWith('here') ? (
 					<Text style={[styles.roleText, name.startsWith('here') && styles.textHere]}>{`@${name}`}</Text>
 				) : (
-					<Text style={styles.title}>{name}</Text>
+					<View style={styles.channelWrapper}>
+						<Text style={styles.title}>{name}</Text>
+						{checkVoiceStatus &&
+							<Text style={styles.channelBusyText}>
+								({t('busy')})
+							</Text>
+						}
+					</View>
+					
 				)}
 			</View>
 			<Text style={styles.subText}>{subText}</Text>
