+++ apps/mobile/src/app/screens/home/homedrawer/components/ChannelList/ChannelListItem/index.tsx
@@ -11,8 +11,7 @@ import { channelsActions, getStoreAsync, selectIsUnreadChannelById, selectVoiceC
 import { ChannelStatusEnum, ChannelThreads, IChannel } from '@mezon/utils';
 import { DrawerActions, useNavigation } from '@react-navigation/native';
 import { ChannelType } from 'mezon-js';
-import React, { useEffect, useMemo, useRef } from 'react';
-import { useTranslation } from 'react-i18next';
+import React, { useEffect, useRef } from 'react';
 import { ActivityIndicator, Linking, Text, TouchableOpacity, View } from 'react-native';
 import { useSelector } from 'react-redux';
 import { linkGoogleMeet } from '../../../../../../utils/helpers';
@@ -41,7 +40,6 @@ export const ChannelListItem = React.memo((props: IChannelListItemProps) => {
 	const voiceChannelMember = useSelector(selectVoiceChannelMembersByChannelId(props?.data?.channel_id));
 	const timeoutRef = useRef<any>();
 	const navigation = useNavigation();
-	const { t } = useTranslation(['clan']);
 
 	useEffect(() => {
 		return () => {
@@ -75,13 +73,6 @@ export const ChannelListItem = React.memo((props: IChannelListItemProps) => {
 		}
 	};
 
-	const checkVoiceStatus = useMemo(() => {
-		if (props?.data?.channel_id !== undefined && voiceChannelMember && props?.data?.type === ChannelType.CHANNEL_TYPE_VOICE) {
-			return voiceChannelMember.length >= 2;
-		}
-		return false;
-	}, [voiceChannelMember, props?.data?.channel_id, props?.data?.type]);
-
 	return (
 		<View>
 			<TouchableOpacity
@@ -109,12 +100,6 @@ export const ChannelListItem = React.memo((props: IChannelListItemProps) => {
 					<Text style={[styles.channelListItemTitle, isUnRead && styles.channelListItemTitleActive]} numberOfLines={1}>
 						{props.data.channel_label}
 					</Text>
-
-					{checkVoiceStatus && (
-						<Text style={styles.channelBusyText}>
-							({t('busy')})
-						</Text>
-					)}
 				</View>
 				{props?.data?.type === ChannelType.CHANNEL_TYPE_VOICE && props?.data?.status === StatusVoiceChannel.No_Active && (
 					<ActivityIndicator color={themeValue.white} />
